separated ParallelContext from parallel.h
