+++ src/parallel/parallel.cpp
@@ -4,7 +4,7 @@
 */
 
 #include <pch.h>
-#include <lm/parallel.h>
+#include <lm/parallelcontext.h>
 
 LM_NAMESPACE_BEGIN(LM_NAMESPACE::parallel)
